requirements.txt
ultralytics>=8.0.0
gradio>=3.38
pillow
numpy
opencv-python-headless
